Restructure code, use Rayon, allow outputs to be chunked
This is a large commit that adds 'chunking' functionality, at the cost
of some complexity. This improves performance significantly. It
splits out the eventually large 'split' module into submodules, each
of which has a different concern.

Specifically, the --chunk-size and --total-rows arguments are now respected.
This required some rethinking about how splits were written to disk. The
pattern now is that we have a 'SplitWriter' for each Split, and the
'main' loop simply sends the row to the relevant SplitWriter. The
SplitWriter handles distributing that to its chunks in a round-robin
fashion. For most combinations of arguments we know how many splits to
create in advance (either using the --total-rows flag, or because we've
been given splits as a fixed number of rows). In this case we spawn one
thread per chunk. If the user just specifies splits as proportions
without the --total-rows flag then we create 2 chunks. The writers for
these chunks will switch to a new file when they reach the chunk size
limit.