Don't try to determine compresssion from input
This only works for files, not for stdin, because we can't seek back to
the start of stdin after checking if we determine that there is no
compression, and therefore end up missing out some of the header.

Instead, require the user to explicitly specify flags if uncompressed
input or output is required.